refactor: 자동차 이름 빈 값 오류 함수를 carNamesErrors.js파일로 이동
